refactor(app): split score screen into focused files
