+++ app/src/main/java/com/clefrun/app/TopOverlayBar.kt
@@ -1,0 +1,135 @@
+package com.clefrun.app
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.layout.statusBarsPadding
+import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.outlined.Refresh
+import androidx.compose.material.icons.outlined.Tune
+import androidx.compose.material3.Icon
+import androidx.compose.material3.Surface
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.text.SpanStyle
+import androidx.compose.ui.text.buildAnnotatedString
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.withStyle
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import com.clefrun.app.ui.theme.Charcoal
+import com.clefrun.app.ui.theme.Paper
+import com.clefrun.app.ui.theme.WarmAccent
+
+@Composable
+internal fun TopOverlayBar(
+    onNewClick: () -> Unit,
+    onOptionsClick: () -> Unit,
+) {
+    Row(
+        modifier = Modifier
+            .fillMaxWidth()
+            .statusBarsPadding()
+            .padding(horizontal = 16.dp, vertical = 12.dp),
+        horizontalArrangement = Arrangement.SpaceBetween,
+        verticalAlignment = Alignment.CenterVertically
+    ) {
+        ClefRunLogo()
+
+        Row(
+            horizontalArrangement = Arrangement.spacedBy(12.dp),
+            verticalAlignment = Alignment.CenterVertically
+        ) {
+            NewExerciseButton(onClick = onNewClick)
+            OptionsButton(onClick = onOptionsClick)
+        }
+    }
+}
+
+@Composable
+private fun ClefRunLogo() {
+    Text(
+        text = buildAnnotatedString {
+            withStyle(
+                style = SpanStyle(
+                    color = Charcoal,
+                    fontWeight = FontWeight.Medium
+                )
+            ) {
+                append("Clef")
+            }
+            withStyle(
+                style = SpanStyle(
+                    color = WarmAccent,
+                    fontWeight = FontWeight.Medium
+                )
+            ) {
+                append("Run")
+            }
+        },
+        fontSize = 26.sp
+    )
+}
+
+@Composable
+private fun NewExerciseButton(
+    onClick: () -> Unit,
+) {
+    Surface(
+        onClick = onClick,
+        shape = RoundedCornerShape(999.dp),
+        color = Paper,
+        contentColor = Charcoal,
+        tonalElevation = 0.dp,
+        shadowElevation = 8.dp
+    ) {
+        Row(
+            modifier = Modifier.padding(horizontal = 18.dp, vertical = 12.dp),
+            verticalAlignment = Alignment.CenterVertically,
+            horizontalArrangement = Arrangement.spacedBy(10.dp)
+        ) {
+            Icon(
+                imageVector = Icons.Outlined.Refresh,
+                contentDescription = "New exercise",
+                tint = Charcoal
+            )
+            Text(
+                text = "New",
+                fontSize = 16.sp,
+                fontWeight = FontWeight.Medium
+            )
+        }
+    }
+}
+
+@Composable
+private fun OptionsButton(
+    onClick: () -> Unit,
+) {
+    Surface(
+        onClick = onClick,
+        shape = CircleShape,
+        color = Paper,
+        contentColor = Charcoal,
+        tonalElevation = 0.dp,
+        shadowElevation = 8.dp
+    ) {
+        Box(
+            modifier = Modifier.size(52.dp),
+            contentAlignment = Alignment.Center
+        ) {
+            Icon(
+                imageVector = Icons.Outlined.Tune,
+                contentDescription = "Exercise options",
+                tint = Charcoal
+            )
+        }
+    }
+}
